RK4

+++ internal/pkg/reviews/repo/reviews_repository.go
@@ -3,9 +3,11 @@ package reviews
 import (
 	"2024_2_ThereWillBeName/internal/models"
 	"2024_2_ThereWillBeName/internal/pkg/dblogger"
+	"2024_2_ThereWillBeName/internal/pkg/outbox"
 	"context"
 	"database/sql"
 	"fmt"
+	"log"
 
 	_ "github.com/lib/pq"
 )
@@ -37,6 +39,13 @@ func (r *ReviewRepository) CreateReview(ctx context.Context, review models.Revie
 		return models.GetReview{}, fmt.Errorf("failed to retrieve created review details: %w", err)
 	}
 
+	payload := fmt.Sprintf(`{"place_id": %d, "user_id": %d}`, review.PlaceID, review.UserID)
+
+	err = outbox.InsertOutboxRecord(ctx, r.db, "review_created", payload)
+	if err != nil {
+		log.Printf("failed to insert outbox record")
+	}
+
 	return createdReview, nil
 }
 
@@ -59,24 +68,36 @@ func (r *ReviewRepository) UpdateReview(ctx context.Context, review models.Revie
 		return fmt.Errorf("no rows were updated: %w", models.ErrNotFound)
 	}
 
+	payload := fmt.Sprintf(`{"place_id": %d, "user_id": %d}`, review.PlaceID, review.UserID)
+
+	err = outbox.InsertOutboxRecord(ctx, r.db, "review_updated", payload)
+
+	if err != nil {
+		return fmt.Errorf("failed to insert outbox record: %w", models.ErrInternal)
+	}
+
 	return nil
 }
 
 func (r *ReviewRepository) DeleteReview(ctx context.Context, reviewID uint) error {
-	query := `DELETE FROM review WHERE id = $1`
+	var placeID, userID int
 
-	result, err := r.db.ExecContext(ctx, query, reviewID)
+	query := `DELETE FROM review WHERE id = $1 RETURNING place_id, user_id`
+
+	err := r.db.QueryRowContext(ctx, query, reviewID).Scan(&placeID, &userID)
 	if err != nil {
+		if err == sql.ErrNoRows {
+			return fmt.Errorf("review not found: %w", models.ErrNotFound)
+		}
 		return fmt.Errorf("failed to delete review: %w", models.ErrInternal)
 	}
 
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return fmt.Errorf("failed to retrieve rows affected: %w", models.ErrInternal)
-	}
+	payload := fmt.Sprintf(`{"place_id": %d, "user_id": %d}`, placeID, userID)
 
-	if rowsAffected == 0 {
-		return fmt.Errorf("review not found: %w", models.ErrNotFound)
+	err = outbox.InsertOutboxRecord(ctx, r.db, "review_deleted", payload)
+
+	if err != nil {
+		return fmt.Errorf("failed to insert outbox record: %w", err)
 	}
 
 	return nil
